// src/gen-ts/src/models/CrupdateExpenseMoney.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface CrupdateExpenseMoney
 */
export interface CrupdateExpenseMoney {
  /**
   *
   * @type {string}
   * @memberof CrupdateExpenseMoney
   */
  comment?: string
  /**
   *
   * @type {number}
   * @memberof CrupdateExpenseMoney
   */
  amount?: number
  /**
   *
   * @type {string}
   * @memberof CrupdateExpenseMoney
   */
  description?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateExpenseMoney
   */
  id?: string
  /**
   * ID of the job this expense is related to
   * @type {string}
   * @memberof CrupdateExpenseMoney
   */
  job_id?: string
}

/**
 * Check if a given object implements the CrupdateExpenseMoney interface.
 */
export function instanceOfCrupdateExpenseMoney(value: object): value is CrupdateExpenseMoney {
  return true
}

export function CrupdateExpenseMoneyFromJSON(json: any): CrupdateExpenseMoney {
  return CrupdateExpenseMoneyFromJSONTyped(json, false)
}

export function CrupdateExpenseMoneyFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CrupdateExpenseMoney {
  if (json == null) {
    return json
  }
  return {
    comment: json['comment'] == null ? undefined : json['comment'],
    amount: json['amount'] == null ? undefined : json['amount'],
    description: json['description'] == null ? undefined : json['description'],
    id: json['id'] == null ? undefined : json['id'],
    job_id: json['job_id'] == null ? undefined : json['job_id'],
  }
}

export function CrupdateExpenseMoneyToJSON(json: any): CrupdateExpenseMoney {
  return CrupdateExpenseMoneyToJSONTyped(json, false)
}

export function CrupdateExpenseMoneyToJSONTyped(
  value?: CrupdateExpenseMoney | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    comment: value['comment'],
    amount: value['amount'],
    description: value['description'],
    id: value['id'],
    job_id: value['job_id'],
  }
}