// src/gen-ts/src/models/Job.ts
export interface Job {
    /**
     * 
     * @type {string}
     * @memberof Job
     */
    comment?: string;
    /**
     * 
     * @type {Date}
     * @memberof Job
     */
    created_at?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Job
     */
    updated_at?: Date;
    /**
     * 
     * @type {AuditUser}
     * @memberof Job
     */
    created_by?: AuditUser;
    /**
     * 
     * @type {AuditUser}
     * @memberof Job
     */
    updated_by?: AuditUser;
    /**
     * 
     * @type {string}
     * @memberof Job
     */
    id?: string;
    /**
     * 
     * @type {CrupdateCompany}
     * @memberof Job
     */
    company?: CrupdateCompany;
    /**
     * 
     * @type {string}
     * @memberof Job
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof Job
     */
    contract_signature_date?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Job
     */
    start_date?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Job
     */
    end_date?: Date;
    /**
     * 
     * @type {JobStatus}
     * @memberof Job
     */
    status?: JobStatus;
}