// src/gen-ts/src/models/CrupdateWarehouse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface CrupdateWarehouse
 */
export interface CrupdateWarehouse {
  /**
   *
   * @type {string}
   * @memberof CrupdateWarehouse
   */
  comment?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateWarehouse
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateWarehouse
   */
  name?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateWarehouse
   */
  description?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateWarehouse
   */
  job_id?: string
}

/**
 * Check if a given object implements the CrupdateWarehouse interface.
 */
export function instanceOfCrupdateWarehouse(value: object): value is CrupdateWarehouse {
  return true
}

export function CrupdateWarehouseFromJSON(json: any): CrupdateWarehouse {
  return CrupdateWarehouseFromJSONTyped(json, false)
}

export function CrupdateWarehouseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CrupdateWarehouse {
  if (json == null) {
    return json
  }
  return {
    comment: json['comment'] == null ? undefined : json['comment'],
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name'],
    description: json['description'] == null ? undefined : json['description'],
    job_id: json['job_id'] == null ? undefined : json['job_id'],
  }
}

export function CrupdateWarehouseToJSON(json: any): CrupdateWarehouse {
  return CrupdateWarehouseToJSONTyped(json, false)
}

export function CrupdateWarehouseToJSONTyped(
  value?: CrupdateWarehouse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    comment: value['comment'],
    id: value['id'],
    name: value['name'],
    description: value['description'],
    job_id: value['job_id'],
  }
}